apps/analyzer/pyproject.toml
[project]
name = "analyzer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "deep-translator>=1.11.4",
    "fastapi>=0.116.1",
    "lingua-language-detector>=2.1.1",
    "torch>=2.7.1",
    "transformers>=4.53.2",
    "uvicorn>=0.35.0",
]
venv = "/Users/arthur/Documents/Dev/smart-moderation-ai/apps/analyzer/.venv/bin/python"
